+++ rtx_package/src/package/pool/latex_ch4_sectioning_and_toc.rs
@@ -49,7 +49,7 @@ LoadDefinitions!(outer_stomach, outer_state, {
     sub[ gullet, (type_tokens, level_arg, ignore3, ignore4, ignore5, ignore6, flag), state ] {
       let stype = type_tokens.to_string();
       let level = level_arg.to_string();
-      let level_int = if level.is_empty() { 0 } else { level.parse::<i32>()? };
+      let level_int = if level.is_empty() { 0 } else { level.parse::<i64>()? };
       let ctr = match state.lookup_value(&s!("counter_for_{}",stype)) {
         Some(v) => v.to_string(),
         None => stype
